New Developments and testing main UI with real local db * Fix note detail window hide/close: now always closing window after   saving/updating note * Fix note list not allowing re-selecting same note after note details   is closed (edge case when we only have one note in db) * Add main menu * Implement (basic) export/import encryption key

+++ service/noteServiceRepository.go
@@ -33,8 +33,9 @@ type NoteServiceRepositoryImpl struct {
 func NewNoteServiceRepository(
 	dbPath string,
 	bucket string,
+	resetDB bool,
 ) (NoteServiceRepository, error) {
-	db, err := openDBConnection(dbPath)
+	db, err := openDBConnection(dbPath, resetDB)
 	if err != nil {
 		return nil, err
 	}
@@ -45,17 +46,19 @@ func NewNoteServiceRepository(
 	}, nil
 }
 
-func openDBConnection(dbPath string) (*nutsdb.DB, error) {
+func openDBConnection(dbPath string, resetDB bool) (*nutsdb.DB, error) {
 	opt := nutsdb.DefaultOptions
 
-	files, _ := ioutil.ReadDir(dbPath)
-	for _, f := range files {
-		name := f.Name()
-		if name != "" {
-			fmt.Println(dbPath + "/" + name)
-			err := os.RemoveAll(dbPath + "/" + name)
-			if err != nil {
-				return nil, err
+	if resetDB {
+		files, _ := ioutil.ReadDir(dbPath)
+		for _, f := range files {
+			name := f.Name()
+			if name != "" {
+				fmt.Println(dbPath + "/" + name)
+				err := os.RemoveAll(dbPath + "/" + name)
+				if err != nil {
+					return nil, err
+				}
 			}
 		}
 	}
